Move all vacuum code into driver
