+++ store/sqlite/sqlite.go
@@ -1,6 +1,7 @@
 package sqlite
 
 import (
+	"context"
 	"database/sql"
 
 	"github.com/usememos/memos/store"
@@ -14,4 +15,51 @@ func NewDriver(db *sql.DB) store.Driver {
 	return &Driver{
 		db: db,
 	}
+}
+
+func (d *Driver) Vacuum(ctx context.Context) error {
+	tx, err := d.db.BeginTx(ctx, nil)
+	if err != nil {
+		return err
+	}
+	defer tx.Rollback()
+
+	if err := vacuumImpl(ctx, tx); err != nil {
+		return err
+	}
+
+	if err := tx.Commit(); err != nil {
+		return err
+	}
+
+	// Vacuum sqlite database file size after deleting resource.
+	if _, err := d.db.Exec("VACUUM"); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func vacuumImpl(ctx context.Context, tx *sql.Tx) error {
+	if err := vacuumMemo(ctx, tx); err != nil {
+		return err
+	}
+	if err := vacuumResource(ctx, tx); err != nil {
+		return err
+	}
+	if err := vacuumUserSetting(ctx, tx); err != nil {
+		return err
+	}
+	if err := vacuumMemoOrganizer(ctx, tx); err != nil {
+		return err
+	}
+	if err := vacuumMemoRelations(ctx, tx); err != nil {
+		return err
+	}
+	if err := vacuumTag(ctx, tx); err != nil {
+		// Prevent revive warning.
+		return err
+	}
+
+	return nil
 }
