+++ store/sqlite/tag.go
@@ -2,6 +2,7 @@ package sqlite
 
 import (
 	"context"
+	"database/sql"
 	"strings"
 
 	"github.com/usememos/memos/store"
@@ -71,5 +72,24 @@ func (d *Driver) DeleteTag(ctx context.Context, delete *store.DeleteTag) error {
 	if _, err = result.RowsAffected(); err != nil {
 		return err
 	}
+	return nil
+}
+
+func vacuumTag(ctx context.Context, tx *sql.Tx) error {
+	stmt := `
+	DELETE FROM 
+		tag 
+	WHERE 
+		creator_id NOT IN (
+			SELECT 
+				id 
+			FROM 
+				user
+		)`
+	_, err := tx.ExecContext(ctx, stmt)
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
