+++ store/sqlite/user.go
@@ -168,5 +168,11 @@ func (d *Driver) DeleteUser(ctx context.Context, delete *store.DeleteUser) error
 	if _, err := result.RowsAffected(); err != nil {
 		return err
 	}
+
+	if err := d.Vacuum(ctx); err != nil {
+		// Prevent linter warning.
+		return err
+	}
+
 	return nil
 }
