+++ store/tag.go
@@ -2,7 +2,6 @@ package store
 
 import (
 	"context"
-	"database/sql"
 )
 
 type Tag struct {
@@ -29,23 +28,4 @@ func (s *Store) ListTags(ctx context.Context, find *FindTag) ([]*Tag, error) {
 
 func (s *Store) DeleteTag(ctx context.Context, delete *DeleteTag) error {
 	return s.driver.DeleteTag(ctx, delete)
-}
-
-func vacuumTag(ctx context.Context, tx *sql.Tx) error {
-	stmt := `
-	DELETE FROM 
-		tag 
-	WHERE 
-		creator_id NOT IN (
-			SELECT 
-				id 
-			FROM 
-				user
-		)`
-	_, err := tx.ExecContext(ctx, stmt)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
