+++ store/user_setting.go
@@ -2,7 +2,6 @@ package store
 
 import (
 	"context"
-	"database/sql"
 
 	storepb "github.com/usememos/memos/proto/gen/store"
 )
@@ -124,23 +123,4 @@ func (s *Store) GetUserAccessTokens(ctx context.Context, userID int32) ([]*store
 
 	accessTokensUserSetting := userSetting.GetAccessTokens()
 	return accessTokensUserSetting.AccessTokens, nil
-}
-
-func vacuumUserSetting(ctx context.Context, tx *sql.Tx) error {
-	stmt := `
-	DELETE FROM 
-		user_setting 
-	WHERE 
-		user_id NOT IN (
-			SELECT 
-				id 
-			FROM 
-				user
-		)`
-	_, err := tx.ExecContext(ctx, stmt)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
